An omission in the right-associative section
The Unary operator is right-associative as we can see by the following example: `var x = -~0b10;`
`~0b10 == 0b11111111111111111111111111111101 == -3` then taking the negative of it we get +3.

Not sure how to interpret the left-associativity of the above operation.  However if we interpret it to mean that the `-` is applied first then `-0b10 == -2 == 0b11111111111111111111111111111110` then apply the bit not operator we get a +1-- not what we get from the evaluation of `-~0b10`!  Which means that unary operators is right-associative: `-~x` is evaluated as `-(~x)`.

Therefore, I propose that we add unary operators as part of the right-associative section.

Thank you for your consideration